plate folding to platonic solids

+++ source/pele/hs_wca.h
@@ -241,7 +241,7 @@ public:
  */
 class HS_WCANeighborList : public SimplePairwiseNeighborList< HS_WCA_interaction > {
 public:
-    HS_WCANeighborList(Array<long int> & ilist, double eps, double sca, Array<double> radii)
+    HS_WCANeighborList(Array<size_t> & ilist, double eps, double sca, Array<double> radii)
         :  SimplePairwiseNeighborList< HS_WCA_interaction > (
                 std::make_shared<HS_WCA_interaction>(eps, sca, radii), ilist)
     {}
